test(desktop): split useAsyncData specs
