+++ apps/desktop/src/composables/__tests__/useAsyncData.test.ts
@@ -1,513 +1,78 @@
 import { useAsyncData } from "@tracepilot/ui";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { describe, expect, it, vi } from "vitest";
 import { nextTick } from "vue";
+import { setupUseAsyncDataTest } from "./useAsyncData/setup";
 
-describe("useAsyncData", () => {
-  beforeEach(() => {
-    vi.useFakeTimers();
+setupUseAsyncDataTest();
+
+describe("useAsyncData basic execution", () => {
+  it("should initialize with default state", () => {
+    const { data, loading, error } = useAsyncData(async () => "test");
+
+    expect(data.value).toBeNull();
+    expect(loading.value).toBe(false);
+    expect(error.value).toBeNull();
   });
 
-  afterEach(() => {
-    vi.restoreAllMocks();
-    vi.useRealTimers();
+  it("should initialize with provided initialData", () => {
+    const { data } = useAsyncData(async () => "test", {
+      initialData: "initial",
+    });
+
+    expect(data.value).toBe("initial");
   });
 
-  describe("Basic Execution", () => {
-    it("should initialize with default state", () => {
-      const { data, loading, error } = useAsyncData(async () => "test");
-
-      expect(data.value).toBeNull();
-      expect(loading.value).toBe(false);
-      expect(error.value).toBeNull();
+  it("should set loading to true during execution", async () => {
+    const asyncFn = vi.fn(async () => {
+      await new Promise((resolve) => setTimeout(resolve, 100));
+      return "result";
     });
 
-    it("should initialize with provided initialData", () => {
-      const { data } = useAsyncData(async () => "test", {
-        initialData: "initial",
-      });
+    const { loading, execute } = useAsyncData(asyncFn);
 
-      expect(data.value).toBe("initial");
-    });
+    const promise = execute();
+    await nextTick();
 
-    it("should set loading to true during execution", async () => {
-      const asyncFn = vi.fn(async () => {
-        await new Promise((resolve) => setTimeout(resolve, 100));
-        return "result";
-      });
+    expect(loading.value).toBe(true);
 
-      const { loading, execute } = useAsyncData(asyncFn);
+    vi.advanceTimersByTime(100);
+    await promise;
 
-      const promise = execute();
-      await nextTick();
-
-      expect(loading.value).toBe(true);
-
-      vi.advanceTimersByTime(100);
-      await promise;
-
-      expect(loading.value).toBe(false);
-    });
-
-    it("should set data on successful execution", async () => {
-      const asyncFn = vi.fn(async () => "success");
-      const { data, execute } = useAsyncData(asyncFn);
-
-      await execute();
-
-      expect(data.value).toBe("success");
-      expect(asyncFn).toHaveBeenCalledTimes(1);
-    });
-
-    it("should pass parameters to async function", async () => {
-      const asyncFn = vi.fn(async (a: number, b: string) => `${a}-${b}`);
-      const { data, execute } = useAsyncData(asyncFn);
-
-      await execute(42, "test");
-
-      expect(data.value).toBe("42-test");
-      expect(asyncFn).toHaveBeenCalledWith(42, "test");
-    });
-
-    it("should clear error on successful execution", async () => {
-      const asyncFn = vi
-        .fn()
-        .mockRejectedValueOnce(new Error("fail"))
-        .mockResolvedValueOnce("success");
-
-      const { error, execute } = useAsyncData(asyncFn);
-
-      await execute();
-      expect(error.value).not.toBeNull();
-
-      await execute();
-      expect(error.value).toBeNull();
-    });
+    expect(loading.value).toBe(false);
   });
 
-  describe("Error Handling", () => {
-    it("should set error on failed execution", async () => {
-      const asyncFn = vi.fn(async () => {
-        throw new Error("Test error");
-      });
+  it("should set data on successful execution", async () => {
+    const asyncFn = vi.fn(async () => "success");
+    const { data, execute } = useAsyncData(asyncFn);
 
-      const { error, execute } = useAsyncData(asyncFn);
+    await execute();
 
-      await execute();
-
-      expect(error.value).toContain("Test error");
-    });
-
-    it("should not set data on failed execution", async () => {
-      const asyncFn = vi.fn(async () => {
-        throw new Error("fail");
-      });
-
-      const { data, execute } = useAsyncData(asyncFn);
-
-      await execute();
-
-      expect(data.value).toBeNull();
-    });
-
-    it("should use custom onError handler", async () => {
-      const asyncFn = vi.fn(async () => {
-        throw new Error("Custom error");
-      });
-
-      const onError = vi.fn((e: unknown) => `Custom: ${(e as Error).message}`);
-      const { error, execute } = useAsyncData(asyncFn, { onError });
-
-      await execute();
-
-      expect(onError).toHaveBeenCalled();
-      expect(error.value).toBe("Custom: Custom error");
-    });
-
-    it("should handle non-Error objects", async () => {
-      const asyncFn = vi.fn(async () => {
-        throw "string error";
-      });
-
-      const { error, execute } = useAsyncData(asyncFn);
-
-      await execute();
-
-      expect(error.value).toBeTruthy();
-    });
-
-    it("should set loading to false after error", async () => {
-      const asyncFn = vi.fn(async () => {
-        throw new Error("fail");
-      });
-
-      const { loading, execute } = useAsyncData(asyncFn);
-
-      await execute();
-
-      expect(loading.value).toBe(false);
-    });
+    expect(data.value).toBe("success");
+    expect(asyncFn).toHaveBeenCalledTimes(1);
   });
 
-  describe("Stale Request Prevention", () => {
-    it("should ignore results from superseded requests", async () => {
-      let resolveFirst!: (value: string) => void;
-      let resolveSecond!: (value: string) => void;
+  it("should pass parameters to async function", async () => {
+    const asyncFn = vi.fn(async (a: number, b: string) => `${a}-${b}`);
+    const { data, execute } = useAsyncData(asyncFn);
 
-      const firstPromise = new Promise<string>((resolve) => {
-        resolveFirst = resolve;
-      });
-      const secondPromise = new Promise<string>((resolve) => {
-        resolveSecond = resolve;
-      });
+    await execute(42, "test");
 
-      const asyncFn = vi.fn().mockReturnValueOnce(firstPromise).mockReturnValueOnce(secondPromise);
-
-      const { data, execute } = useAsyncData(asyncFn);
-
-      const first = execute();
-      const second = execute();
-
-      resolveSecond?.("second");
-      await second;
-
-      expect(data.value).toBe("second");
-
-      resolveFirst?.("first");
-      await first;
-
-      expect(data.value).toBe("second");
-    });
-
-    it("should ignore errors from superseded requests", async () => {
-      let rejectFirst!: (error: Error) => void;
-      let resolveSecond!: (value: string) => void;
-
-      const firstPromise = new Promise<string>((_, reject) => {
-        rejectFirst = reject;
-      });
-      const secondPromise = new Promise<string>((resolve) => {
-        resolveSecond = resolve;
-      });
-
-      const asyncFn = vi.fn().mockReturnValueOnce(firstPromise).mockReturnValueOnce(secondPromise);
-
-      const { data, error, execute } = useAsyncData(asyncFn);
-
-      const first = execute();
-      const second = execute();
-
-      resolveSecond?.("success");
-      await second;
-
-      expect(data.value).toBe("success");
-      expect(error.value).toBeNull();
-
-      rejectFirst?.(new Error("old error"));
-      await first.catch(() => {});
-
-      expect(error.value).toBeNull();
-    });
-
-    it("should not update loading state for stale requests", async () => {
-      let resolveFirst!: (value: string) => void;
-      let resolveSecond!: (value: string) => void;
-
-      const firstPromise = new Promise<string>((resolve) => {
-        resolveFirst = resolve;
-      });
-      const secondPromise = new Promise<string>((resolve) => {
-        resolveSecond = resolve;
-      });
-
-      const asyncFn = vi.fn().mockReturnValueOnce(firstPromise).mockReturnValueOnce(secondPromise);
-
-      const { loading, execute } = useAsyncData(asyncFn);
-
-      execute();
-      execute();
-
-      expect(loading.value).toBe(true);
-
-      resolveSecond?.("second");
-      await nextTick();
-
-      expect(loading.value).toBe(false);
-
-      resolveFirst?.("first");
-      await nextTick();
-
-      expect(loading.value).toBe(false);
-    });
+    expect(data.value).toBe("42-test");
+    expect(asyncFn).toHaveBeenCalledWith(42, "test");
   });
 
-  describe("Immediate Execution", () => {
-    it("should execute immediately when immediate=true", async () => {
-      const asyncFn = vi.fn(async () => "immediate");
-      const { data, loading } = useAsyncData(asyncFn, { immediate: true });
-
-      expect(loading.value).toBe(true);
-
-      await vi.runAllTimersAsync();
-
-      expect(data.value).toBe("immediate");
-      expect(asyncFn).toHaveBeenCalledTimes(1);
-    });
-
-    it("should not execute immediately when immediate=false", () => {
-      const asyncFn = vi.fn(async () => "data");
-      useAsyncData(asyncFn, { immediate: false });
-
-      expect(asyncFn).not.toHaveBeenCalled();
-    });
-
-    it("should execute immediately with default immediate value", () => {
-      const asyncFn = vi.fn(async () => "data");
-      useAsyncData(asyncFn);
-
-      expect(asyncFn).not.toHaveBeenCalled();
-    });
-  });
-
-  describe("Success Callback", () => {
-    it("should call onSuccess when execution succeeds", async () => {
-      const onSuccess = vi.fn();
-      const asyncFn = vi.fn(async () => "success");
-
-      const { execute } = useAsyncData(asyncFn, { onSuccess });
-
-      await execute();
-
-      expect(onSuccess).toHaveBeenCalledWith("success");
-      expect(onSuccess).toHaveBeenCalledTimes(1);
-    });
-
-    it("should not call onSuccess when execution fails", async () => {
-      const onSuccess = vi.fn();
-      const asyncFn = vi.fn(async () => {
-        throw new Error("fail");
-      });
-
-      const { execute } = useAsyncData(asyncFn, { onSuccess });
-
-      await execute();
-
-      expect(onSuccess).not.toHaveBeenCalled();
-    });
-
-    it("should not call onSuccess for stale requests", async () => {
-      const onSuccess = vi.fn();
-      let resolveFirst!: (value: string) => void;
-      let resolveSecond!: (value: string) => void;
-
-      const firstPromise = new Promise<string>((resolve) => {
-        resolveFirst = resolve;
-      });
-      const secondPromise = new Promise<string>((resolve) => {
-        resolveSecond = resolve;
-      });
-
-      const asyncFn = vi.fn().mockReturnValueOnce(firstPromise).mockReturnValueOnce(secondPromise);
-
-      const { execute } = useAsyncData(asyncFn, { onSuccess });
-
-      execute();
-      execute();
-
-      resolveSecond?.("second");
-      await nextTick();
-
-      expect(onSuccess).toHaveBeenCalledTimes(1);
-      expect(onSuccess).toHaveBeenCalledWith("second");
-
-      resolveFirst?.("first");
-      await nextTick();
-
-      expect(onSuccess).toHaveBeenCalledTimes(1);
-    });
-  });
-
-  describe("Refresh", () => {
-    it("should re-execute with last parameters", async () => {
-      const asyncFn = vi.fn(async (x: number) => x * 2);
-      const { data, execute, refresh } = useAsyncData(asyncFn);
-
-      await execute(5);
-      expect(data.value).toBe(10);
-
-      await refresh();
-      expect(data.value).toBe(10);
-      expect(asyncFn).toHaveBeenCalledTimes(2);
-      expect(asyncFn).toHaveBeenNthCalledWith(1, 5);
-      expect(asyncFn).toHaveBeenNthCalledWith(2, 5);
-    });
-
-    it("should be a no-op if execute was never called", async () => {
-      const asyncFn = vi.fn(async () => "data");
-      const { refresh } = useAsyncData(asyncFn);
-
-      await refresh();
-
-      expect(asyncFn).not.toHaveBeenCalled();
-    });
-
-    it("should use most recent parameters", async () => {
-      const asyncFn = vi.fn(async (x: number) => x);
-      const { data, execute, refresh } = useAsyncData(asyncFn);
-
-      await execute(1);
-      await execute(2);
-      await execute(3);
-
-      await refresh();
-
-      expect(data.value).toBe(3);
-      expect(asyncFn).toHaveBeenLastCalledWith(3);
-    });
-  });
-
-  describe("Clear Error", () => {
-    it("should clear error message", async () => {
-      const asyncFn = vi.fn().mockRejectedValue(new Error("fail"));
-      const { error, execute, clearError } = useAsyncData(asyncFn);
-
-      await execute();
-      expect(error.value).not.toBeNull();
-
-      clearError();
-      expect(error.value).toBeNull();
-    });
-
-    it("should not affect data or loading", async () => {
-      const asyncFn = vi.fn().mockRejectedValue(new Error("fail"));
-      const { data, loading, execute, clearError } = useAsyncData(asyncFn, {
-        initialData: "initial",
-      });
-
-      await execute();
-
-      clearError();
-
-      expect(data.value).toBe("initial");
-      expect(loading.value).toBe(false);
-    });
-  });
-
-  describe("Reset", () => {
-    it("should reset all state to initial values", async () => {
-      const asyncFn = vi.fn(async (x: number) => x * 2);
-      const { data, loading, error, execute, reset } = useAsyncData(asyncFn, {
-        initialData: 42,
-      });
-
-      await execute(5);
-      expect(data.value).toBe(10);
-
-      reset();
-
-      expect(data.value).toBe(42);
-      expect(loading.value).toBe(false);
-      expect(error.value).toBeNull();
-    });
-
-    it("should invalidate in-flight requests", async () => {
-      let resolveAsync!: (value: string) => void;
-      const asyncPromise = new Promise<string>((resolve) => {
-        resolveAsync = resolve;
-      });
-
-      const asyncFn = vi.fn().mockReturnValue(asyncPromise);
-      const { data, loading, execute, reset } = useAsyncData(asyncFn);
-
-      const promise = execute();
-      expect(loading.value).toBe(true);
-
-      reset();
-      expect(loading.value).toBe(false);
-
-      resolveAsync?.("result");
-      await promise;
-
-      expect(data.value).toBeNull();
-    });
-
-    it("should reset lastParams so refresh is a no-op", async () => {
-      const asyncFn = vi.fn(async (x: number) => x);
-      const { execute, reset, refresh } = useAsyncData(asyncFn);
-
-      await execute(42);
-      reset();
-      await refresh();
-
-      expect(asyncFn).toHaveBeenCalledTimes(1);
-    });
-  });
-
-  describe("Type Safety", () => {
-    it("should infer data type from async function return type", async () => {
-      const asyncFn = async (): Promise<{ id: number; name: string }> => ({
-        id: 1,
-        name: "test",
-      });
-
-      const { data, execute } = useAsyncData(asyncFn);
-
-      await execute();
-
-      expect(data.value?.id).toBe(1);
-      expect(data.value?.name).toBe("test");
-    });
-
-    it("should infer parameter types from async function", async () => {
-      const asyncFn = async (id: number, name: string) => ({ id, name });
-
-      const { execute } = useAsyncData(asyncFn);
-
-      await execute(1, "test");
-    });
-  });
-
-  describe("Edge Cases", () => {
-    it("should handle null or undefined results", async () => {
-      const asyncFn = vi.fn(async () => null);
-      const { data, execute } = useAsyncData(asyncFn);
-
-      await execute();
-
-      expect(data.value).toBeNull();
-    });
-
-    it("should handle async function that returns undefined", async () => {
-      const asyncFn = vi.fn(async () => undefined);
-      const { data, execute } = useAsyncData(asyncFn);
-
-      await execute();
-
-      expect(data.value).toBeUndefined();
-    });
-
-    it("should handle rapid execute calls", async () => {
-      const asyncFn = vi.fn(async (x: number) => x);
-      const { data, execute } = useAsyncData(asyncFn);
-
-      execute(1);
-      execute(2);
-      execute(3);
-      execute(4);
-      const final = execute(5);
-
-      await final;
-
-      expect(data.value).toBe(5);
-    });
-
-    it("should handle execute with zero parameters", async () => {
-      const asyncFn = vi.fn(async () => "no params");
-      const { data, execute } = useAsyncData(asyncFn);
-
-      await execute();
-
-      expect(data.value).toBe("no params");
-      expect(asyncFn).toHaveBeenCalledWith();
-    });
+  it("should clear error on successful execution", async () => {
+    const asyncFn = vi
+      .fn()
+      .mockRejectedValueOnce(new Error("fail"))
+      .mockResolvedValueOnce("success");
+
+    const { error, execute } = useAsyncData(asyncFn);
+
+    await execute();
+    expect(error.value).not.toBeNull();
+
+    await execute();
+    expect(error.value).toBeNull();
   });
 });
