+++ apps/desktop/src/composables/__tests__/useAsyncData/setup.ts
@@ -1,0 +1,12 @@
+import { afterEach, beforeEach, vi } from "vitest";
+
+export function setupUseAsyncDataTest(): void {
+  beforeEach(() => {
+    vi.useFakeTimers();
+  });
+
+  afterEach(() => {
+    vi.restoreAllMocks();
+    vi.useRealTimers();
+  });
+}
